data-ingestion/requirements.txt
psycopg2-binary
requests
schedule
PyYAML
